requirements.txt
# GUI Framework
PySide6>=6.6.0

# Excel/CSV Verarbeitung
openpyxl>=3.1.2
pandas>=2.1.0

# PDF Export
reportlab>=4.0.0

# Visualisierung
plotly>=5.18.0

# OpenAI Integration
openai>=1.10.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-qt>=4.2.0

# Code Quality
black>=23.0.0
flake8>=6.1.0
mypy>=1.5.0
